// src/dto/admission.dto.ts
class AdmissionDTO {
  firstName: string;
  lastName: string;
  middleName: string;
  email: string;
  phoneNumber: string;
  dateOfBirth: string;
  gender: string;
  nationality: string;
  postalAddress: string;
  maritalStatus: string;
  languages: string[];
  programChoice: string;
  reference: string;
  addressOfReference: string;
  phoneOfReference: string;
  emailOfReference: string;
  parentName: string;
  parentPhoneNumber: string;
  parentEmail: string;
  parentAddress: string;
  sponsor: string;
  qualification: string[];
  institution: string[];
  yearOfCompletion: string[];
  country: string[];
  startDate: string[];
  endDate: string[];
}

export default AdmissionDTO;